Add remove button and functionality
