--- src/components/AddButton.js
@@ -1,44 +1,0 @@
-import { useDispatch } from 'react-redux';
-import React, { useState } from 'react';
-import { v4 as uuidv4 } from 'uuid';
-import { addBook } from '../Redux/books/booksSlice';
-
-const AddBtn = () => {
-  const [title, setTitle] = useState('');
-  const [author, setAuthor] = useState('');
-  const dispatch = useDispatch();
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    const newBook = {
-      id: uuidv4(),
-      title,
-      author,
-    };
-    dispatch(addBook(newBook));
-    setTitle(' ');
-    setAuthor(' ');
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <input
-        type="text"
-        placeholder="Book title"
-        value={title}
-        onChange={(e) => setTitle(e.target.value)}
-      />
-      <input
-        type="text"
-        placeholder="Author"
-        value={author}
-        onChange={(e) => setAuthor(e.target.value)}
-      />
-      <button type="submit">
-        Add Book
-      </button>
-    </form>
-  );
-};
-
-export default AddBtn;
